Refactor code; add css modules
