+++ src/components/CryptoCurrency/CryptoCurrency.module.css
@@ -1,0 +1,3 @@
+.CryptoCurrency {
+  border: 1px solid gray;
+}
